fixed video reloads on posts
